Step3,4 - 블랙잭 딜러, 베팅 (#149)
* refactor: Step2 #145 피드백 반영

- 일급 컬렉션이 불변이 될 수 있도록 방어적 복사를 한다.
- 모든 fun 의 리턴 타입을 명시한다.

* feat: 딜러 구현

* feat: 승부에 따라 다른 profit 이 나오도록 함

* feat: 포장되지 않은 원시값 제거

+++ src/test/kotlin/domain/player/BetAmountTest.kt
@@ -22,6 +22,6 @@ internal class BetAmountTest {
     @Test
     fun times() {
         assertThat(BetAmount(1000) * 1.5)
-            .isEqualTo((1500).toDouble())
+            .isEqualTo(1500.0)
     }
 }
